+++ packages/engine/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @flatkit/engine
 
+## 0.5.0
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @flatkit/types@0.5.0
+
 ## 0.4.0
 
 ### Minor Changes