+++ packages/player/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @flatkit/player
 
+## 0.5.0
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @flatkit/types@0.5.0
+  - @flatkit/engine@0.5.0
+
 ## 0.4.0
 
 ### Patch Changes