+++ packages/types/CHANGELOG.md
@@ -1,5 +1,7 @@
 # @flatkit/types
 
+## 0.5.0
+
 ## 0.4.0
 
 ### Minor Changes